add "Perform Updates" button
